1) Converted the following classes to use BigInteger instead of int:   1.a) Eratosthenes   1.b) PrimeFactory   1.c) FactorizationFactory   1.d) CommonPolynomial   1.e) AlgebraicPolynomial   1.f) RationalPolynomial   1.g) SkewSymmetricPolynomial   1.h) BitMatrix   1.i) BitVector   1.j) Relation   1.k) FactorPair   1.l) FactorCollection   1.m) MainForm

+++ GNFS-Winforms/MainForm.cs
@@ -55,7 +55,7 @@ namespace GNFS_Winforms
 			n = BigInteger.Parse(tbN.Text);
 			degree = int.Parse(tbDegree.Text);
 
-			IEnumerable<int> primes = PrimeFactory.GetPrimes(10000);
+			IEnumerable<BigInteger> primes = PrimeFactory.GetPrimes(10000);
 			BigInteger baseM = CommonPolynomial.SuggestPolynomialBase(n, degree, primes);
 
 			tbBase.Text = baseM.ToString();
